Finish encode, add haversine, debug indexing

+++ algo/src/max_dtw.rs
@@ -1,14 +1,14 @@
 use crate::rest::Point;
 
-pub fn max_dtw(ta: &[Point], tb: &[Point]) -> i32 {
+pub fn max_dtw(ta: &[Point], tb: &[Point]) -> f64 {
     match (ta, tb) {
-        ([], []) => 0,
+        ([], []) => 0.0,
         ([.., a], [.., b]) => a.distance(b).max(q(ta, tb)),
-        _ => i32::MAX,
+        _ => f64::MAX,
     }
 }
 
-fn q(ta: &[Point], tb: &[Point]) -> i32 {
+fn q(ta: &[Point], tb: &[Point]) -> f64 {
     max_dtw(except_last(&ta), except_last(tb))
         .min(max_dtw(except_last(ta), &tb))
         .min(max_dtw(&ta, except_last(tb)))
